test: invariant + full-flow integration tests + Slither static analysis
5 new TODO items checked off (Phase 1 Tests + Static Analysis):

Tests added (75/75 passing across 9 suites):
- test/invariant/VaultInvariants.t.sol — 4 invariants via VaultHandler fuzzer:
    nonNegativeState, sharesBackedByAssets, totalAssetsContainsLocalBalance,
    treasurySharesMonotonic
  Documents one DESIGN NOTE: principalTracked can exceed totalAssets if agent
  reports cross-chain assets lower after harvest — by design, vault stays solvent.
- test/integration/MirrorFlow.t.sol — full lifecycle in one test:
    pool init → 2-user deposit (80k USDC) → real V4 swap → cross-chain yield
    report (4k USDC) → time warp + harvest → partial redeem. Fee math validated
    (569.86 USDC treasury shares for ~3.99k extra yield). Runs in 3.6s.

Static analysis:
- Installed Slither 0.11.5
- .slither.config.json: filter_paths for lib/ + test/ + script/, exclude
  low-level-calls (V4 hook needs them), naming-convention, timestamp early-returns,
  solc-version (deps), assembly (deps), too-many-digits (deps).
- Fixed 2 legit findings:
    1. MirrorHook._handleEvent CEI: lastDispatchTime now set BEFORE _dispatchToAllSisters
       (prevents mailbox re-entry from re-triggering dispatch in same block)
    2. MirrorVault.PerformanceFeePaid event: treasury parameter now indexed

Code quality:
- forge fmt clean across all files
- forge inspect storage-layout snapshots saved for Hook/Vault/Factory/Treasury/Relayer
  (used for upgrade-safety diffs in Phase 6)

Deferred to Phase 6 audit prep:
- Mythril (slow symbolic execution; less actionable than Slither solo)
- Bankr aeon-vuln-scanner (needs Bankr account setup)

Test summary:
  MirrorFactoryTest:        8
  MirrorVaultTest:         15
  MockHyperlaneMailboxTest: 6
  RelayerTest:             16
  TreasuryTest:            12
  ForkBaseTest:             9
  HookCallbackTest:         4
  MirrorFlowTest:           1   ← new
  VaultInvariantsTest:      4   ← new
  TOTAL:                   75

+++ packages/contracts/src/interfaces/IPyth.sol
@@ -5,27 +5,27 @@ pragma solidity 0.8.26;
 /// @dev Contract addresses: https://docs.pyth.network/price-feeds/contract-addresses
 interface IPyth {
     struct Price {
-        int64  price;       // Price scaled by 10^expo
-        uint64 conf;        // Confidence interval
-        int32  expo;        // Negative exponent (e.g. -8 means price is in units of 1e-8)
-        uint   publishTime; // Unix timestamp of price
+        int64 price; // Price scaled by 10^expo
+        uint64 conf; // Confidence interval
+        int32 expo; // Negative exponent (e.g. -8 means price is in units of 1e-8)
+        uint256 publishTime; // Unix timestamp of price
     }
 
     struct PriceFeed {
         bytes32 id;
-        Price   price;
-        Price   emaPrice;
+        Price price;
+        Price emaPrice;
     }
 
     /// @notice Get the current price for a feed. Reverts if price is stale.
     function getPrice(bytes32 id) external view returns (Price memory price);
 
     /// @notice Get the price, reverting if older than `age` seconds
-    function getPriceNoOlderThan(bytes32 id, uint age) external view returns (Price memory price);
+    function getPriceNoOlderThan(bytes32 id, uint256 age) external view returns (Price memory price);
 
     /// @notice Push fresh price data on-chain (call before reading price on-chain)
     function updatePriceFeeds(bytes[] calldata updateData) external payable;
 
     /// @notice ETH fee required to call updatePriceFeeds
-    function getUpdateFee(bytes[] calldata updateData) external view returns (uint feeAmount);
+    function getUpdateFee(bytes[] calldata updateData) external view returns (uint256 feeAmount);
 }
